Adding Scheduler.c, updated thread_list.c, and thread_object.c
* Added some basic structures and functions to shceduler.c
* Updated the findNodeAtIndex function, only visits each node once
instead of visiting a node for every ticket the node has. (subtracts
node's tickets from total instead of one at a time)
* removed some whitepsaces in the removeThread method, there was a space
between a function call and it's arguments, and some whitepsaces in the
pointer strings
* added inclusion of thread_object.h to thread_object.c

Note: I have not tested the code, I am just uploading it so the changes
aren't lost.